Rename `config` dir to `utils`
